+++ src/pages/Dashboard/index.tsx
@@ -1,7 +1,6 @@
 import { AddCircleFillIcon } from "@assets/svgs";
 import { useState } from "react";
 import AddModal from "./components/AddModal";
-import CancelModal from "@components/CancelModal";
 import MonitoringCard from "./components/MonitoringCard";
 import SearchBar from "@components/SearchBar";
 import ClientNotFound from "./components/ClientNotFound";
@@ -11,7 +10,6 @@ import usePostSearchClient from "@api/hooks/clients/usePostSearchClient";
 
 const DashboardPage: React.FC = () => {
   const [isAddModalOpen, setIsAddModalOpen] = useState(false);
-  const [isModalOpen, setIsModalOpen] = useState(false);
   const [searchText, setSearchText] = useState("");
   const [searchQuery, setSearchQuery] = useState("");
 
@@ -27,18 +25,8 @@ const DashboardPage: React.FC = () => {
     setIsAddModalOpen(true);
   };
 
-  const handleModalClose = () => {
-    setIsModalOpen(false);
-  };
-
   const handleClose = () => {
     setIsAddModalOpen(false);
-    setIsModalOpen(true);
-  };
-
-  const handleCancel = () => {
-    setIsAddModalOpen(true);
-    setIsModalOpen(false);
   };
 
   const handleSubmit = () => {
@@ -68,7 +56,7 @@ const DashboardPage: React.FC = () => {
         <ClientNotFound handleAddModalOpen={handleAddModalOpen} />
       ) : (
         <>
-          <div className="mb-4 mt-[63px] flex w-[1048px] justify-between">
+          <div className="mb-[42px] mt-[63px] flex w-[1048px] justify-between">
             <SearchBar
               value={searchText}
               onChange={handleChangeSearchText}
@@ -102,18 +90,6 @@ const DashboardPage: React.FC = () => {
       )}
       {isAddModalOpen && (
         <AddModal onClose={handleClose} onSubmit={handleSubmit} />
-      )}
-      {isModalOpen && (
-        <CancelModal
-          onClose={handleModalClose}
-          handleCancel={handleCancel}
-          headingText="정말 중단하시겠어요?"
-          bodyText={`작성하던 모든 기록은 지워지며 
-              이후 복구가 불가능해요.
-              창을 정말 닫으시겠어요?`}
-          closeButtonText="창 닫기"
-          cancelButtonText="취소"
-        />
       )}
     </div>
   );
